+++ src/main/java/com/raoulvdberge/refinedpipes/RefinedPipesBlocks.java
@@ -18,6 +18,10 @@ public class RefinedPipesBlocks {
     public static final FluidPipeBlock IMPROVED_FLUID_PIPE = null;
     @ObjectHolder(RefinedPipes.ID + ":advanced_fluid_pipe")
     public static final FluidPipeBlock ADVANCED_FLUID_PIPE = null;
+    @ObjectHolder(RefinedPipes.ID + ":elite_fluid_pipe")
+    public static final FluidPipeBlock ELITE_FLUID_PIPE = null;
+    @ObjectHolder(RefinedPipes.ID + ":ultimate_fluid_pipe")
+    public static final FluidPipeBlock ULTIMATE_FLUID_PIPE = null;
     @ObjectHolder(RefinedPipes.ID + ":basic_energy_pipe")
     public static final EnergyPipeBlock BASIC_ENERGY_PIPE = null;
     @ObjectHolder(RefinedPipes.ID + ":improved_energy_pipe")