+++ src/main/java/com/raoulvdberge/refinedpipes/network/pipe/fluid/FluidPipeType.java
@@ -9,7 +9,9 @@ import net.minecraft.util.ResourceLocation;
 public enum FluidPipeType {
     BASIC(1),
     IMPROVED(2),
-    ADVANCED(3);
+    ADVANCED(3),
+    ELITE(4),
+    ULTIMATE(5);
 
     private final int tier;
 
@@ -29,6 +31,10 @@ public enum FluidPipeType {
                 return RefinedPipesTileEntities.IMPROVED_FLUID_PIPE;
             case ADVANCED:
                 return RefinedPipesTileEntities.ADVANCED_FLUID_PIPE;
+            case ELITE:
+                return RefinedPipesTileEntities.ELITE_FLUID_PIPE;
+            case ULTIMATE:
+                return RefinedPipesTileEntities.ULTIMATE_FLUID_PIPE;
             default:
                 throw new RuntimeException("?");
         }
@@ -42,6 +48,10 @@ public enum FluidPipeType {
                 return RefinedPipes.SERVER_CONFIG.getImprovedFluidPipe().getCapacity();
             case ADVANCED:
                 return RefinedPipes.SERVER_CONFIG.getAdvancedFluidPipe().getCapacity();
+            case ELITE:
+                return RefinedPipes.SERVER_CONFIG.getEliteFluidPipe().getCapacity();
+            case ULTIMATE:
+                return RefinedPipes.SERVER_CONFIG.getUltimateFluidPipe().getCapacity();
             default:
                 throw new RuntimeException("?");
         }
@@ -55,6 +65,10 @@ public enum FluidPipeType {
                 return RefinedPipes.SERVER_CONFIG.getImprovedFluidPipe().getTransferRate();
             case ADVANCED:
                 return RefinedPipes.SERVER_CONFIG.getAdvancedFluidPipe().getTransferRate();
+            case ELITE:
+                return RefinedPipes.SERVER_CONFIG.getEliteFluidPipe().getTransferRate();
+            case ULTIMATE:
+                return RefinedPipes.SERVER_CONFIG.getUltimateFluidPipe().getTransferRate();
             default:
                 throw new RuntimeException("?");
         }
@@ -68,6 +82,10 @@ public enum FluidPipeType {
                 return new ResourceLocation(RefinedPipes.ID, "improved_fluid_pipe");
             case ADVANCED:
                 return new ResourceLocation(RefinedPipes.ID, "advanced_fluid_pipe");
+            case ELITE:
+                return new ResourceLocation(RefinedPipes.ID, "elite_fluid_pipe");
+            case ULTIMATE:
+                return new ResourceLocation(RefinedPipes.ID, "ultimate_fluid_pipe");
             default:
                 throw new RuntimeException("?");
         }
@@ -81,6 +99,10 @@ public enum FluidPipeType {
                 return new ResourceLocation(RefinedPipes.ID, "improved_fluid_network");
             case ADVANCED:
                 return new ResourceLocation(RefinedPipes.ID, "advanced_fluid_network");
+            case ELITE:
+                return new ResourceLocation(RefinedPipes.ID, "elite_fluid_network");
+            case ULTIMATE:
+                return new ResourceLocation(RefinedPipes.ID, "ultimate_fluid_network");
             default:
                 throw new RuntimeException("?");
         }